pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "difference calculator"
authors = ["Aleksey Chudinov <alekseychudinov@gmail.com>"]
repository = "https://github.com/alekseychudinov/python-project-lvl2"
readme = "README.md"
classifiers = [
 "Programming Language :: Python :: 3.10"
]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
